% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AlignmentAligner.R
\name{align_alignments}
\alias{align_alignments}
\title{Score the agreement between two alignments}
\usage{
align_alignments(ref.fa, aln.fa)
}
\arguments{
\item{ref.fa}{Reference against which to score}

\item{aln.fa}{Alignment to score against reference}
}
\value{
Results matrix with one column per sequence and the following rows
\itemize{
 \item ColumnMatch
 \item NonGap
 \item Cys
 \item Match
 \item Gap(con)
 \item Insertion
 \item Deletion
 \item Substitution
 \item FinalMatch
 \item Score
}
}
\description{
Score the agreement between two alignments
}
\examples{
data(ref)
data(aln)
results <- align_alignments(ref,aln)
}

