\name{AlignStat-package}
\alias{AlignStat-package}
\alias{AlignStat}
\docType{package}
\title{
\packageTitle{AlignStat}
}
\description{
\packageDescription{AlignStat}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{AlignStat}
\packageIndices{AlignStat}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{AlignStat}

Maintainer: \packageMaintainer{AlignStat}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%% ~~ simple examples of the most important functions ~~
}
