% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{comparison_alignment}
\alias{comparison_alignment}
\title{Comparison Alignment}
\format{A data frame with sequences in columns and amino acid in rows"comparison_alignment"}
\usage{
comparison_alignment
}
\description{
An example alignment
}
\keyword{datasets}

