% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Plots.R
\name{match_summary_plot}
\alias{match_summary_plot}
\title{Plot match proportion + Cys}
\usage{
match_summary_plot(results, ref)
}
\arguments{
\item{results}{As produced by align_alignments}

\item{ref}{Reference alignment}
}
\description{
Plot match proportion + Cys
}
\examples{
data(ref)
data(aln)
res_list <- align_alignments(ref,aln)
match_summary_plot(res_list$results,ref)
}

