% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{plot_alignment_heatmap}
\alias{plot_alignment_heatmap}
\title{Plot a heatmap of the column identities between two multiple sequence alignments}
\usage{
plot_alignment_heatmap(x, display = TRUE)
}
\arguments{
\item{x}{an object of type "pairwise alignment comparison" (typically the summary file generated by compare_alignments)}

\item{display}{display this plot (default = TRUE)}
}
\description{
Plot a heatmap of the column identities between two multiple sequence alignments
}
\note{
The `plot_alignment_heatmap` displays the similarity between each pairwise column comparison for the reference and comparison multiple sequence alignments. Colour density is determined by the proportion of identical character matches (M) between the columns, normalised to the number of characters that are not merely conserved gaps (not G). The (M) and (not G) values are generated by the `compare_alignments` function as the result list items `$results` and `$means` respectively. This gives a representation of which columns are well agreed upon by the MSAs, and which columns are split by one MSA relative to the other.
}
\examples{
data(reference_alignment)
data(comparison_alignment)
PAC <- compare_alignments(reference_alignment,comparison_alignment)
plot_alignment_heatmap(PAC)

}

