% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{plot_category_proportions}
\alias{plot_category_proportions}
\title{Plot the different causes of column dissimilarity between two multiple sequence alignments}
\usage{
plot_category_proportions(x, stack = FALSE, display = TRUE)
}
\arguments{
\item{x}{an object of type "pairwise alignment comparison" (typically the summary file generated by compare_alignments)}

\item{stack}{stacked area plot in stead of line plot (deault = FALSE)}

\item{display}{display this plot (default = TRUE)}
}
\description{
Plot the different causes of column dissimilarity between two multiple sequence alignments
}
\note{
This function generates a detailed breakdown of the differences between the multiple sequence alignments. For all characters that are neither identical residues, nor conserved gaps, the relative proportions of insertions (I), deletions (D) and substitutions (S) is plotted.
}
\examples{
data(reference_alignment)
data(comparison_alignment)
PAC <- compare_alignments(reference_alignment,comparison_alignment)
plot_category_proportions(PAC, stack=TRUE)

}

