% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{plot_match_summary}
\alias{plot_match_summary}
\title{Plot a summary of column similarity between two multiple sequence alignments}
\usage{
plot_match_summary(x, cys = FALSE, display = TRUE)
}
\arguments{
\item{x}{an object of type "pairwise alignment comparison" (typically the summary file generated by compare_alignments)}

\item{cys}{additionally show the cysteine abundance for each column (default = FALSE)}

\item{display}{display this plot (default = TRUE)}
}
\description{
Plot a summary of column similarity between two multiple sequence alignments
}
\note{
This function generates a plot that summarises the agreement between the two multiple sequence alignments. It plots the proportion of identical character matches (M) for each column as a proportion of the characters that are not merely conserved gaps (not G). The overall average proportion of identical characters that are not conserved gaps is overlaid as a percentage. For alignments of cysteine-rich proteins, it can additionally plot the cysteine abundance for each column to indicate columns containing conserved cysteines (`cys=TRUE`).
}
\examples{
data(reference_alignment)
data(comparison_alignment)
PAC <- compare_alignments(reference_alignment,comparison_alignment)
plot_match_summary(PAC, cys=TRUE)

}

