% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Plots.R
\name{proportion_cys_plot}
\alias{proportion_cys_plot}
\title{Plot match proportion + Cys}
\usage{
proportion_cys_plot(results, ref)
}
\arguments{
\item{results}{As produced by align_alignments}

\item{ref}{Reference alignment}
}
\description{
Plot match proportion + Cys
}
\examples{
data(ref)
data(aln)
results <- align_alignments(ref,aln)
proportion_cys_plot(results,ref)
}

