% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reference_alignment}
\alias{reference_alignment}
\title{Reference Alignment}
\format{
A data frame with sequences in columns and amino acid in rows
}
\usage{
reference_alignment
}
\description{
An example alignment
}
\keyword{datasets}
