% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Plots.R
\name{alignment_heatmap}
\alias{alignment_heatmap}
\title{Alignment heatmap}
\usage{
alignment_heatmap(results, means, aln, ref, display = TRUE)
}
\arguments{
\item{results}{As produced by align_alignments}

\item{aln}{Alignment to align to ref}

\item{ref}{Reference alignment}

\item{whether}{to display plot}
}
\description{
Alignment heatmap
}
\examples{
data(ref)
data(aln)
res_list <- align_alignments(ref,aln)
alignment_heatmap(res_list$results,res_list$means,aln,ref)
}

