% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Plots.R
\name{alignment_heatmap}
\alias{alignment_heatmap}
\title{Alignment heatmap}
\usage{
alignment_heatmap(results, aln, ref)
}
\arguments{
\item{results}{As produced by align_alignments}

\item{aln}{Alignment to align to ref}

\item{ref}{Reference alignment}
}
\description{
Alignment heatmap
}
\examples{
data(ref)
data(aln)
results <- align_alignments(ref,aln)
alignment_heatmap(results)
}

