\name{AlikeAlignmentAligner-package}
\alias{AlikeAlignmentAligner-package}
\alias{AlikeAlignmentAligner}
\docType{package}
\title{
\packageTitle{AlikeAlignmentAligner}
}
\description{
\packageDescription{AlikeAlignmentAligner}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{AlikeAlignmentAligner}
\packageIndices{AlikeAlignmentAligner}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{AlikeAlignmentAligner}

Maintainer: \packageMaintainer{AlikeAlignmentAligner}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%% ~~ simple examples of the most important functions ~~
}
