% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Plots.R
\name{category_proportions_plot}
\alias{category_proportions_plot}
\title{Category proportions plot}
\usage{
category_proportions_plot(results)
}
\arguments{
\item{results}{As produced by align_alignments}
}
\description{
Category proportions plot
}
\examples{
data(ref)
data(aln)
res_list <- align_alignments(ref,aln)
category_proportions_plot(res_list$results)
}

