% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_alignments.R
\name{compare_alignments}
\alias{compare_alignments}
\title{Compare alternative multiple sequence alignments}
\usage{
compare_alignments(ref, com)
}
\arguments{
\item{ref}{The reference MSA (in fasta format)}

\item{com}{The MSA to compare (in fasta format)}
}
\value{
Generates an object of class "pairwise alignment comparison" (PAC), providing the optimal alignment of alignments and comparison of the differences between them. The details of the PAC output components are as follows:
\itemize{
 \item {Columnmatch}   {The column of the comparison alignment with the highest final match score}
 \item {Nongap}        {The proportion of characters that are not gaps}
 \item {Cys}           {The proportion of cysteines (relevant for cysteine rich proteins)}
 \item {Match}         {The proportion of characters that are identical between alignments}
 \item {Gapcon}        {The proportion of characters that are conserved gaps}
 \item {Insertion}     {The proportion of characters that are a gap in the reference, but are a residue in the comparison alignment}
 \item {Deletion}      {The proportion of characters that are a residue in the reference, but a gap in the comparison alignment}
 \item {Substitution}  {The proportion of characters that are one residue in the reference, but a non-homologous residue in the comparison alignment}
 \item {Finalmatch}    {The proportion of characters that match as a proportion of those that are not conserved gaps (Match/not_Gapcon)}
}
}
\description{
Compare alternative multiple sequence alignments
}
\note{
This function aligns two multiple sequence alignments (MSA) against one another. The alternative alignments must contain the same sequences in the same order. The function will classify any similarities and differences between the two MSAs. It produces the "pairwise alignment comparison" object required as the first step any other package functions.

This The `compare_alignments` function first checks that the MSAs are alternative alignments of the same sequences. The function labels each character in the alignment with its occurrence number in the sequence (e.g. to distinguish between the first and second cysteines of a sequence). It then compares the two MSAs to determine which columns are the closest matches between the ref and com MSAs. Each pairwise column comparison is stored as the `$means` value of the output. From this matrix, the comparison alignment column with the highest final match to each reference alignment column is used to calculate further statistics for the `$results` value of the output. The overall Finalmatch score for the whole comparison is output as the `$score` value.
}
\examples{
data(reference_alignment)
data(comparison_alignment)
PAC <- compare_alignments(reference_alignment,comparison_alignment)

}

